+++ BoardGameProject/src/main/java/com/team4/controllers/DeckController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.team4.model.classes.uno.Deck;
+import com.team4.model.Deck;
 import com.team4.repositories.DeckRepository;
 
 @RestController
