+++ BoardGameProject/src/main/java/com/team4/controllers/PieceController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.team4.model.abstrct.Piece;
+import com.team4.model.Piece;
 import com.team4.repositories.PieceRepository;
 
 @RestController
