+++ BoardGameProject/src/main/java/com/team4/model/Chess.java
@@ -1,0 +1,59 @@
+package com.team4.model;
+
+import java.util.List;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.JoinTable;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+
+@Entity
+public class Chess extends Game {
+	
+	@OneToMany(fetch = FetchType.EAGER, cascade = CascadeType.MERGE)
+	@JoinTable(
+			name = "GamePlayer",
+			joinColumns = @JoinColumn(name = "gameID"),
+			inverseJoinColumns = @JoinColumn(name = "playerID"))
+	private List<Player> players;
+	
+	@OneToMany(fetch = FetchType.EAGER, cascade = CascadeType.MERGE)
+	private List<Rule> rules;
+	
+
+	@OneToOne
+	private Board board;
+	
+	public Chess() { }
+
+	public Chess(String type, List<Player> players, List<Rule> rules) {
+		super();
+		board = new Board();
+		this.players = players;
+		this.rules = rules;
+	}
+	
+	public List<Player> getPlayers() {
+		return players;
+	}
+	
+	public void setPlayers(List<Player> players) {
+		this.players = players;
+	}
+	
+	public List<Rule> getRules() {
+		return rules;
+	}
+	
+	public void setRules(List<Rule> rules) {
+		this.rules = rules;
+	}
+	
+	public Board getBoard() { return board; }
+	
+	public void setBoard(Board board) { this.board = board; }
+	
+}
