+++ BoardGameProject/src/main/java/com/team4/model/Hand.java
@@ -1,4 +1,4 @@
-package com.team4.model.classes.uno;
+package com.team4.model;
 
 import java.util.List;
