+++ BoardGameProject/src/main/java/com/team4/model/Pawn.java
@@ -1,23 +1,21 @@
-package com.team4.model.classes.chess.pieces;
+package com.team4.model;
 
 import java.util.ArrayList;
 import java.util.List;
-
-import com.team4.model.abstrct.Piece;
-import com.team4.model.classes.chess.Space;
 
 import jakarta.persistence.Entity;
 
 
 @Entity
-public class King extends Piece {
+public class Pawn extends Piece {
 	
-	public King() {
+	
+	public Pawn() {
 		
 	}
 	
-    public King(Color color, Space currentSpace) {
-        super(color, Piece.Type.KING ,currentSpace);
+    public Pawn(Color color, Space currentSpace) {
+        super(color, Piece.Type.PAWN, currentSpace);
     }
 
     @Override
@@ -26,4 +24,5 @@ public class King extends Piece {
         // We gotta add some code here
         return possibleMoves;
     }
+
 }
