+++ BoardGameProject/src/main/java/com/team4/model/Queen.java
@@ -1,12 +1,9 @@
-package com.team4.model.classes.chess.pieces;
+package com.team4.model;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import jakarta.persistence.Entity;
-
-import com.team4.model.abstrct.Piece;
-import com.team4.model.classes.chess.Space;
 
 @Entity
 public class Queen extends Piece {
