+++ BoardGameProject/src/main/java/com/team4/model/UserAccount.java
@@ -1,4 +1,4 @@
-package com.team4.model.classes;
+package com.team4.model;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
