+++ BoardGameProject/src/main/java/com/team4/repositories/BishopRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.chess.pieces.Bishop;
+import com.team4.model.Bishop;
 
 public interface BishopRepository extends JpaRepository<Bishop, Long> {
