+++ BoardGameProject/src/main/java/com/team4/repositories/BoardRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.chess.Board;
+import com.team4.model.Board;
 
 public interface BoardRepository extends JpaRepository<Board, Long> {
