+++ BoardGameProject/src/main/java/com/team4/repositories/CardRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.uno.Card;
+import com.team4.model.Card;
 
 public interface CardRepository extends JpaRepository<Card, Long> {
