+++ BoardGameProject/src/main/java/com/team4/repositories/ChessRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.chess.Chess;
+import com.team4.model.Chess;
 
 public interface ChessRepository extends JpaRepository<Chess, Long> {
