+++ BoardGameProject/src/main/java/com/team4/repositories/DeckRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.uno.Deck;
+import com.team4.model.Deck;
 
 public interface DeckRepository extends JpaRepository<Deck, Long>{
