+++ BoardGameProject/src/main/java/com/team4/repositories/GameRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.abstrct.Game;
+import com.team4.model.Game;
 
 public interface GameRepository extends JpaRepository<Game, Long> {
