+++ BoardGameProject/src/main/java/com/team4/repositories/HandRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.uno.Hand;
+import com.team4.model.Hand;
 
 public interface HandRepository extends JpaRepository<Hand, Long>{
