+++ BoardGameProject/src/main/java/com/team4/repositories/KingRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.chess.pieces.King;
+import com.team4.model.King;
 
 public interface KingRepository extends JpaRepository<King, Long> {
