+++ BoardGameProject/src/main/java/com/team4/repositories/KnightRepository.java
@@ -3,7 +3,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.chess.pieces.Knight;
+import com.team4.model.Knight;
 
 public interface KnightRepository extends JpaRepository<Knight, Long> {
