+++ BoardGameProject/src/main/java/com/team4/repositories/PawnRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.chess.pieces.Pawn;
+import com.team4.model.Pawn;
 
 public interface PawnRepository extends JpaRepository<Pawn, Long> {
