+++ BoardGameProject/src/main/java/com/team4/repositories/PieceRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.abstrct.Piece;
+import com.team4.model.Piece;
 
 public interface PieceRepository extends JpaRepository<Piece, Long>{
