+++ BoardGameProject/src/main/java/com/team4/repositories/PlayerRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.Player;
+import com.team4.model.Player;
 
 public interface PlayerRepository extends JpaRepository<Player, Long> {
