+++ BoardGameProject/src/main/java/com/team4/repositories/QueenRepository.java
@@ -3,7 +3,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.chess.pieces.Queen;
+import com.team4.model.Queen;
 
 public interface QueenRepository extends JpaRepository<Queen, Long> {
