+++ BoardGameProject/src/main/java/com/team4/repositories/RookRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.chess.pieces.Rook;
+import com.team4.model.Rook;
 
 public interface RookRepository extends JpaRepository<Rook, Long> {
