+++ BoardGameProject/src/main/java/com/team4/repositories/RuleRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.Rule;
+import com.team4.model.Rule;
 
 public interface RuleRepository extends JpaRepository<Rule, Long>{
