+++ BoardGameProject/src/main/java/com/team4/repositories/SpaceRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.chess.Space;
+import com.team4.model.Space;
 
 public interface SpaceRepository extends JpaRepository<Space, Long> {
