+++ BoardGameProject/src/main/java/com/team4/repositories/UnoRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.uno.Uno;
+import com.team4.model.Uno;
 
 public interface UnoRepository extends JpaRepository<Uno, Long> {
