+++ BoardGameProject/src/main/java/com/team4/repositories/UserAccountRepository.java
@@ -2,7 +2,7 @@ package com.team4.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team4.model.classes.UserAccount;
+import com.team4.model.UserAccount;
 
 public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
